+++ changelog.md
@@ -8,6 +8,7 @@
 - Update to project template 4.3.3
 - Removed submodules and added the libraries as packages
 - Updated question flow adding social closeness input in the question creation flow instead of similarity and reason
+- Integrated domain data in the question flow
 - Refactored PendingMessagesJob
 
 ## 1.0.10